fix(store): keep Unix free-space math lint-clean

+++ crates/coop-store/src/lib.rs
@@ -689,7 +689,7 @@ fn filesystem_available_bytes(path: &Path) -> StoreResult<u64> {
         return Err(std::io::Error::last_os_error().into());
     }
     let stats = unsafe { stats.assume_init() };
-    Ok((stats.f_bavail as u64).saturating_mul(stats.f_frsize as u64))
+    Ok(stats.f_bavail.saturating_mul(stats.f_frsize))
 }
 
 #[cfg(windows)]
